requirements.txt
discord.py>=2.3.2
python-dotenv>=1.0.0
aiosqlite>=0.19.0
pytest>=8.2.0
pytest-asyncio>=0.23.5
numpy>=2.3.2
profanity-check>=1.0.3
scikit-learn>=1.7.1

